requirements.txt
python-dotenv
boto3
requests
pandas
matplotlib
seaborn
plotly
scikit-learn